+++ src/app.module.ts
@@ -4,6 +4,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { validateEnv } from './config/env.validation';
 import { BrandsModule } from './brands/brands.module';
+import { ClientsModule } from './clients/clients.module';
 import { DatabaseModule } from './database/database.module';
 import { DocumentTypesModule } from './document-types/document-types.module';
 import { HealthModule } from './health/health.module';
@@ -22,6 +23,7 @@ import { ProductsModule } from './products/products.module';
     DatabaseModule,
     HealthModule,
     DocumentTypesModule,
+    ClientsModule,
     BrandsModule,
     ProductsModule,
   ],